Replaced type 'byte' for 'int'
- `byte` is not making a single difference here in perfomance.
- Thus, it just adds a level of boiler-plater that is not satisfying me.
- The decision was to remove it and use `int` instead.

+++ Assets/ChessPieces/Bishop.cs
@@ -2,6 +2,6 @@ using System;
 
 namespace Chess.Assets.ChessPieces;
 
-public class Bishop(char symbol, ConsoleColor color, byte x, byte y) : ChessPiece(symbol, color, x, y)
+public class Bishop(char symbol, ConsoleColor color, int x, int y) : ChessPiece(symbol, color, x, y)
 {
 }
