+++ Assets/ChessPieces/Queen.cs
@@ -2,6 +2,6 @@ using System;
 
 namespace Chess.Assets.ChessPieces;
 
-public class Queen(char symbol, ConsoleColor color, byte x, byte y) : ChessPiece(symbol, color, x, y)
+public class Queen(char symbol, ConsoleColor color, int x, int y) : ChessPiece(symbol, color, x, y)
 {
 }
